+++ src/mako/common/AdapterManager.php
@@ -109,18 +109,6 @@ abstract class AdapterManager
 	/**
 	 * Returns an instance of the chosen adapter configuration.
 	 *
-	 * @deprecated
-	 * @param  string|null $configuration Configuration name
-	 * @return mixed
-	 */
-	public function instance(?string $configuration = null)
-	{
-		return $this->getInstance($configuration);
-	}
-
-	/**
-	 * Returns an instance of the chosen adapter configuration.
-	 *
 	 * @param  string|null $configuration Configuration name
 	 * @return mixed
 	 */